+++ CHANGELOG.md
@@ -8,6 +8,7 @@ All notable changes to this project will be documented in this file, per [the Ke
 
 - Bump WordPress version "tested up to" 6.7.
 - Update littlefoot.js to 4.1.2.
+- Plugin description.
 
 ## [1.2.0] - 2024-03-17